Initial MYSql database
Initial creation of the database. Includes tables as shown in the DataModel.jpg

All 5 tables are populated with data.
For user & scores tables, I included a few test entries to work with.
The scores table includes a date column that automaticaally timestamps when a row is created.  I have tested, and it works.  I included a few test objects set to earlier dates so not all dates are the same.