Got the scores into the GraphResults.js file to show the scores.
Passed the scores from the LoginServlet to the results.jsp. Retrieved
them in a string named scores. Used request.getAttribute() to get them
into the results.jsp, then made an input tag named scores. Retrieved the
string in GraphResults.js to scores, split the string with a comma, then
parsed the values to an integer in "arrScores". Set the labels to the
graph with "count", set the values to "arrScores".